+++ examples/example_codegg.py
@@ -38,12 +38,12 @@ o.seed_elements(13., 67.8, z=-40, radius=2000, number=500,
 
 #%%
 # Adjusting some configuration
-o.set_config('processes:turbulentmixing', True)
-#o.set_config('turbulentmixing:diffusivitymodel', 'windspeed_Sundby1983') # windspeed parameterization for eddy diffusivity
-o.set_config('turbulentmixing:diffusivitymodel', 'environment') # use eddy diffusivity from ocean model
+o.set_config('drift:vertical_mixing', True)
+#o.set_config('vertical_mixing:diffusivitymodel', 'windspeed_Sundby1983') # windspeed parameterization for eddy diffusivity
+o.set_config('vertical_mixing:diffusivitymodel', 'environment') # use eddy diffusivity from ocean model
 #%%
 # Vertical mixing requires fast time step
-o.set_config('turbulentmixing:timestep', 60.) # seconds
+o.set_config('vertical_mixing:timestep', 60.) # seconds
 
 #%%
 # Running model