+++ examples/example_reader_boundary.py
@@ -34,7 +34,7 @@ lons, lats = np.meshgrid(lons, lats)
 o.seed_elements(lons, lats, radius=0, number=2500,
                 time=reader_norkyst.start_time)
 
-o.set_config('processes:turbulentmixing', False)
+o.set_config('drift:vertical_mixing', False)
 
 #%%
 # Running model