+++ examples/example_seed_from_shapefile.py
@@ -20,7 +20,7 @@ o.seed_from_shapefile(o.test_data_folder() +
 o.fallback_values['x_wind'] = -4  # Constant wind drift
 o.fallback_values['y_wind'] = 8
 o.set_config('drift:wind_uncertainty', 4) # Adding some diffusion
-o.set_config('processes:turbulentmixing', False)
+o.set_config('drift:vertical_mixing', False)
 
 #%%
 # Running model