+++ tests/integration/test_models.py
@@ -128,8 +128,8 @@ class TestModels(unittest.TestCase):
         reader_current = reader_netCDF_CF_generic.Reader(o.test_data_folder() +
                 '14Jan2016_NorKyst_z_3d/NorKyst-800m_ZDEPTHS_his_00_3Dsubset.nc')
 
-        reader_landmask = reader_global_landmask.Reader(llcrnrlon=3., llcrnrlat=60.,
-                            urcrnrlon=5., urcrnrlat=63.5)
+        reader_landmask = reader_global_landmask.Reader(
+                extent=[3., 5., 60., 63.5])
 
         o.add_reader([reader_current,reader_landmask])
         o.seed_elements(4.,62.,time=reader_current.start_time)